refactor(lib, docs and example): adjusting code reference and doc before publishing
