+++ lib/index.ts
@@ -4,3 +4,6 @@ export * from './daemon';
 export * from './emitter';
 export * from './event';
 export * from './listener';
+export * from './listeners';
+export * from './broker';
+export * from './utils/errors';
